parcar :updated vehicleService

+++ src/main/java/com/hcl/parkar/dao/VehicleRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import com.hcl.parkar.model.VehicleEntity;
 
 @Repository
-public interface VehicleRepository extends CrudRepository<VehicleEntity, Long> {
+public interface VehicleRepository extends CrudRepository<VehicleEntity, Integer> {
 	
-	Optional<VehicleEntity>findByVehicleNumber(int vehicleNumber);
+	Optional<VehicleEntity> findByVehicleNumber(int vehicleNumber);
 
 }
